+++ gobblin-api/src/main/java/gobblin/writer/initializer/WriterInitializer.java
