+++ gobblin-core/src/main/java/gobblin/writer/initializer/WriterInitializerFactory.java
@@ -21,8 +21,8 @@ import gobblin.configuration.ConfigurationKeys;
 import gobblin.configuration.State;
 import gobblin.source.workunit.WorkUnitStream;
 import gobblin.util.ForkOperatorUtils;
+import gobblin.writer.DataWriterBuilder;
 
-import java.lang.reflect.Method;
 import java.util.List;
 
 import com.google.common.base.Preconditions;
@@ -32,9 +32,6 @@ import com.google.common.collect.Lists;
  * Factory method pattern class provides WriterInitializer based on writer and state.
  */
 public class WriterInitializerFactory {
-  private static final String GET_INITIALIZER_METHOD = "getInitializer";
-  private static final Class<?> GET_INITIALIZER_PARAMS[] = { State.class, WorkUnitStream.class, int.class, int.class };
-
   private static final NoopWriterInitializer NOOP = new NoopWriterInitializer();
 
   /**
@@ -63,25 +60,18 @@ public class WriterInitializerFactory {
     String writerBuilderKey = ForkOperatorUtils.getPropertyNameForBranch(ConfigurationKeys.WRITER_BUILDER_CLASS, branches, branchId);
     String writerBuilderClass = state.getProp(writerBuilderKey, ConfigurationKeys.DEFAULT_WRITER_BUILDER_CLASS);
 
-    Class<?> klazz;
+    DataWriterBuilder dataWriterBuilder;
     try {
-      klazz = Class.forName(writerBuilderClass);
-    } catch (ClassNotFoundException e) {
-      throw new RuntimeException(e);
-    }
-
-    Method method;
-    try {
-      method = klazz.getMethod(GET_INITIALIZER_METHOD, GET_INITIALIZER_PARAMS);
-    } catch (NoSuchMethodException e) {
-      return NOOP;
-    }
-
-    try {
-      Object initializer = method.invoke(klazz.newInstance(), state, workUnits, branches, branchId);
-      return (WriterInitializer) initializer;
+      dataWriterBuilder = (DataWriterBuilder) Class.forName(writerBuilderClass).newInstance();
     } catch (Exception e) {
       throw new RuntimeException(e);
     }
+
+    WriterInitializer initializer = dataWriterBuilder.getInitializer(state, workUnits, branches, branchId);
+    if (initializer != null) {
+      return initializer;
+    }
+
+    return NOOP;
   }
 }
