+++ gobblin-core/src/main/resources/mysql/mysql.properties
@@ -16,7 +16,7 @@
 #
 
 # Source properties - source class to extract data from Mysql Source
-source.class=gobblin.source.jdbc.MysqlSource
+source.class=gobblin.source.extractor.extract.jdbc.MysqlSource
 
 # Source properties
 source.max.number.of.partitions=<1>
