+++ gobblin-core/src/main/resources/sqlserver/sqlserver.properties
@@ -16,7 +16,7 @@
 #
 
 # Source properties - source class to extract data from sqlserver Source
-source.class=gobblin.source.jdbc.SqlServerSource
+source.class=gobblin.source.extractor.extract.jdbc.SqlServerSource
 
 # Source properties
 source.max.number.of.partitions=<1>
