+++ gobblin-modules/gobblin-sql/src/test/resources/converter/user.avsc
@@ -1,0 +1,45 @@
+{
+  "type" : "record",
+  "name" : "TUPLE_0",
+  "fields" : [ {
+    "name" : "userId",
+    "type" : [ "null", "string" ],
+    "doc" : "autogenerated from Pig Field Schema"
+  }, {
+    "name" : "memberId",
+    "type" : [ "null", "long" ],
+    "doc" : "autogenerated from Pig Field Schema"
+  }, {
+    "name" : "businessUnit",
+    "type" : [ "null", "string" ],
+    "doc" : "autogenerated from Pig Field Schema"
+  }, {
+    "name" : "level",
+    "type" : [ "null", "string" ],
+    "doc" : "autogenerated from Pig Field Schema"
+  }, {
+    "name" : "geoRegion",
+    "type" : [ "null", "string" ],
+    "doc" : "autogenerated from Pig Field Schema"
+  }, {
+    "name" : "superRegion",
+    "type" : [ "null", "string" ],
+    "doc" : "autogenerated from Pig Field Schema"
+  }, {
+    "name" : "subRegion",
+    "type" : [ "null", "string" ],
+    "doc" : "autogenerated from Pig Field Schema"
+  }, {
+    "name" : "currency",
+    "type" : [ "null", "string" ],
+    "doc" : "autogenerated from Pig Field Schema"
+  }, {
+    "name" : "segment",
+    "type" : [ "null", "string" ],
+    "doc" : "autogenerated from Pig Field Schema"
+  }, {
+    "name" : "vertical",
+    "type" : [ "null", "string" ],
+    "doc" : "autogenerated from Pig Field Schema"
+  } ]
+}
